services/ingest/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.2
opentelemetry-sdk>=1.24
opentelemetry-exporter-otlp>=1.24
opentelemetry-exporter-otlp-proto-grpc>=1.24
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
opentelemetry-instrumentation-logging>=0.45b0
opentelemetry-instrumentation-system-metrics>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
opentelemetry-semantic-conventions>=0.45b0
